Update Entity classes and Instructor dashboard
Updated constructors of User and its subclasses. Updated the body of the login function in Login class. Added fillData function in Instructor and User and added the attributes of the Instructor. Added isValidUser function to the User class. Added the constructor of the InstructorDashboard class and displayed the name of the Instructor in the dashboard.