Enhancement done to the DBScript and Data
DBScript: Modified the naming convention for one table to match other tables and unified the database case to UPPER case letters.
Data: Added more instances for the tables and started modifying them to suit the proper final state of the database in order to replace the testing version of it.